refactor/contract to domain
