+++ src/domain/integrations/http/jsonplaceholder/jsonplaceholder.interface.ts
