+++ src/domain/integrations/http/jsonplaceholder/jsonplaceholder.response.ts
