+++ src/domain/respositories/index.ts
@@ -1,0 +1,3 @@
+export * from './cache/todo/todo.interface'
+export * from './no_sql/todos/todo.interface'
+export * from './relational/todo/todo.interface'
