+++ src/infra/integrations/index.ts
@@ -2,17 +2,12 @@ export * from './http/client';
 export * from './http/interfaces';
 export * from './http/enum';
 export * from './http/jsonplaceholder/jsonplaceholder';
-export * from './http/jsonplaceholder/jsonplaceholder.response';
-export * from './http/jsonplaceholder/jsonplaceholder.interface';
 export * from './amqp/client';
 export * from './amqp/interface';
 export * from './amqp/type';
 export * from './amqp/producer/producer';
 export * from './amqp/producer/types';
 export * from './amqp/producer/todo/todo';
-export * from './amqp/producer/todo/todo.interface';
-export * from './grpc/notification/notification.interface';
-export * from './amqp/producer/notification/notification.interface';
 export * from './amqp/producer/notification/notification';
 export * from './grpc/grpc.client';
 export * from './grpc/grpc.interface';
