+++ src/infra/repositories/cache/todo/todo.ts
@@ -1,5 +1,6 @@
 import { Todo } from 'src/domain/entities';
-import { CacheClient, ITodoCache } from 'src/infra/repositories';
+import { ITodoCache } from 'src/domain/respositories';
+import { CacheClient } from 'src/infra/repositories';
 
 export class TodoCache implements ITodoCache {
   constructor(private cache: CacheClient) {}
